Fixed flakiness in MasterMaintenanceTest.InverseOffers.
The accept/decline of the inverse offers in the test could arrive after the clock is paused and advanced in the test.  This meant that the test would trigger an allocation and then idle (because the clock is paused).  This is the same problem we fixed in https://issues.apache.org/jira/browse/MESOS-4059 .

Review: https://reviews.apache.org/r/43971/