Ignored subsequent status update in HealthStatusChange tests.
In HealthStatusChange test cases, we launch a task that toggles between
healthy and unhealthy, and will never be killed because no consecutive
health failures occur. We need to ignore subsequent status updates it
is possible to continue to receive status updates before we stop the
driver.

Review: https://reviews.apache.org/r/46307