Fixed the incomplete `TaskStatus` message of docker executor.
When we enable health check for the tasks run in docker container,
their `TaskStatus` messages generated by docker executor may miss
`NetworkInfo` field and agent would fill the host ip as default value.
In this changes, we cache the `NetworkInfo` of the task and reuse it
when generate `TaskStatus` messages which state is `TASK_RUNNING`.

Review: https://reviews.apache.org/r/50103